Render Meshlets (#29)
* feat: Uniform testing

* feat: Draw Meshlets

* chore: Added stricter compiler flags and fixed all warnings

* feat: Draw Cube

* feat: constant buffer instead of root constants

* feat: Packed Buffer

* feat: Testing Framework

* chore: Add noexcept

* feat: Meshlet deletion

+++ core/src/win32/str.cpp
@@ -25,7 +25,7 @@ core::str::wide_to_string(std::wstring_view ws)
 	int size_needed =
 		WideCharToMultiByte(CP_UTF8, 0, ws.data(), (int)ws.size(), nullptr, 0, nullptr, nullptr);
 
-	std::string result(size_needed, 0);
+	std::string result(static_cast<size_t>(size_needed), 0);
 
 	WideCharToMultiByte(
 		CP_UTF8,
